+++ CHANGELOG.md
@@ -1,5 +1,4 @@
 # Product Backlog
-* [Medium] Fix promise created in a handler warning in express-sequelize-session/lib/e4store.js:57 to 61 (maybe fork)
 * [High] Build DELETE /user/attendance API method
 * [High] Bound history in ui.bootstrap.popover (https://angular-ui.github.io/bootstrap/#!#popover)
 * [High] Intrapage links don't work because anchors with ids load asynchronously
@@ -10,6 +9,8 @@
 * [Low] Braintree subscription payments for Teacher Training's multiple payments option
 
 <a name="1.0.7"></a>
+* [Medium] Replace express-sequelize-session with node-connect-pg-simple or connect-session-sequelize
+* Removed cookie-parser as express-session (since 1.5) no longer needs it
 * Updates for node and webpack
 
 <a name="1.0.6"></a>